* Add in documentation for `Sample::from_i*`
* Change to larger integer number for bits per sample large than 16

+++ src/frame/decoder.rs
@@ -4,7 +4,7 @@ use frame::ChannelAssignment;
 //
 // Two channels, left and side (difference) that transforms the difference
 // into the right channel.
-pub fn decode_left_side(buffer: &mut [i32]) {
+pub fn decode_left_side(buffer: &mut [i64]) {
   let block_size = buffer.len() / 2;
 
   for i in 0..block_size {
@@ -20,7 +20,7 @@ pub fn decode_left_side(buffer: &mut [i32]) {
 //
 // Two channels, side (difference) and right that transforms the difference
 // into the left channel.
-pub fn decode_right_side(buffer: &mut [i32]) {
+pub fn decode_right_side(buffer: &mut [i64]) {
   let block_size = buffer.len() / 2;
 
   for i in 0..block_size {
@@ -36,7 +36,7 @@ pub fn decode_right_side(buffer: &mut [i32]) {
 //
 // Two channels, midpoint (average) and side (difference) that transforms
 // the average and difference into the left and right channels.
-pub fn decode_midpoint_side(buffer: &mut [i32]) {
+pub fn decode_midpoint_side(buffer: &mut [i64]) {
   let block_size = buffer.len() / 2;
 
   for i in 0..block_size {
@@ -60,7 +60,7 @@ pub fn decode_midpoint_side(buffer: &mut [i32]) {
 ///   channels.
 /// * `MidpointSide` - decode midpoint and side channels to left and right
 ///   channels.
-pub fn decode(channel_assignment: ChannelAssignment, buffer: &mut [i32]) {
+pub fn decode(channel_assignment: ChannelAssignment, buffer: &mut [i64]) {
   match channel_assignment {
     ChannelAssignment::Independent  => return,
     ChannelAssignment::LeftSide     => decode_left_side(buffer),
